requirements/dev.txt
-r base.txt

pre-commit==2.20.0
identify==2.4.2
django-stubs==1.9.0
mypy==0.931
mock-django==0.6.10
pytest-django==4.4.0
django-debug-toolbar==3.2.4
coverage==6.3.1
gunicorn==20.1.0
